[Feature] Restructure torchrl/objectives (#580)
* restructure torchrl/objectives

* restructure torchrl/objectives

* lint

* fix formatting of file tree in demo notebook
