+++ test/smoke_test.py
@@ -8,4 +8,4 @@ def test_imports():
     from torchrl.envs import Transform, TransformedEnv  # noqa: F401
     from torchrl.envs.gym_like import GymLikeEnv  # noqa: F401
     from torchrl.modules import TensorDictModule  # noqa: F401
-    from torchrl.objectives.costs.common import LossModule  # noqa: F401
+    from torchrl.objectives.common import LossModule  # noqa: F401
