+++ torchrl/objectives/__init__.py
@@ -3,5 +3,17 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .costs import *
-from .returns import *
+from .common import LossModule
+from .ddpg import DDPGLoss
+from .dqn import DQNLoss, DistributionalDQNLoss
+from .ppo import PPOLoss, ClipPPOLoss, KLPENPPOLoss
+from .redq import REDQLoss
+from .sac import SACLoss
+from .utils import (
+    SoftUpdate,
+    HardUpdate,
+    distance_loss,
+    hold_out_params,
+    next_state_value,
+)
+from .value import bellman_max, c_val, dv_val, vtrace, GAE, TDLambdaEstimate, TDEstimate
