--- torchrl/objectives/costs/__init__.py
@@ -1,12 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-#
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
-
-from .common import *
-from .ddpg import *
-from .dqn import *
-from .ppo import *
-from .sac import *
-from .redq import *
-from .utils import *
