+++ torchrl/objectives/ppo.py
@@ -11,11 +11,8 @@ from torch import distributions as d
 
 from torchrl.data.tensordict.tensordict import TensorDictBase, TensorDict
 from torchrl.modules import TensorDictModule
-from ...modules.tensordict_module import ProbabilisticTensorDictModule
-
-__all__ = ["PPOLoss", "ClipPPOLoss", "KLPENPPOLoss"]
-
-from torchrl.objectives.costs.utils import distance_loss
+from torchrl.objectives.utils import distance_loss
+from ..modules.tensordict_module import ProbabilisticTensorDictModule
 from .common import LossModule
