+++ torchrl/objectives/redq.py
@@ -14,14 +14,12 @@ from torch import Tensor
 from torchrl.data.tensordict.tensordict import TensorDictBase, TensorDict
 from torchrl.envs.utils import set_exploration_mode, step_mdp
 from torchrl.modules import TensorDictModule
-from torchrl.objectives.costs.common import LossModule
-from torchrl.objectives.costs.utils import (
+from torchrl.objectives.common import LossModule
+from torchrl.objectives.utils import (
     distance_loss,
     hold_out_params,
     next_state_value as get_next_state_value,
 )
-
-__all__ = ["REDQLoss"]
 
 
 class REDQLoss(LossModule):
