+++ torchrl/objectives/utils.py
@@ -15,8 +15,6 @@ from torchrl.data.tensordict.tensordict import TensorDictBase
 from torchrl.envs.utils import step_mdp
 from torchrl.modules import TensorDictModule
 
-__all__ = ["SoftUpdate", "HardUpdate", "distance_loss", "hold_out_params"]
-
 
 class _context_manager:
     def __init__(self, value=True):
